Added dialog + New items
Added more dialog to the game in the cafeteria as well as the generator room, and the storage closet now contains the code for the master chamber. The game is currently at the point where the player moves from the storage closet into the living quarters, and sees a geashes, it will then scurry off and the player is left to enter the master chamber and use the code to find the weapon.